requirements.txt
Flask==3.1.0
Flask-MySQLdb==2.0.0
Werkzeug==3.1.3
groq==0.24.0
mysql.connector==2.2.9
tree_sitter_language_pack==0.7.3
pytz==2024.2
pillow==11.1.0
pytesseract==0.3.13
